beginning integration
i got the color bars to restrict their colors to rgb, the timer sets and can be started and reset once started. in the morning i need to fix a bug where the random does not display the color initally (it displays on the second click of start). after that the next step is getting the values of the guess and calculating score from the results